+++ Changelog.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Removed
 - Removed the `dry_run` option (https://github.com/zendesk/global_uid/pull/64)
+- Removed `GlobalUid::ServerVariables` module (https://github.com/zendesk/global_uid/pull/66)
 
 ## [3.7.1] - 2020-02-06
 ### Added